Fix some .F's
